refactor: reorganize experiments and routes
Move experiments into folders; PascalCase non-route components; move routes into subfolders; update imports and routes.ts; tests/typecheck pass
