+++ app/experiments/cafe-wall/CafeWall.tsx
@@ -126,4 +126,3 @@ function Field({
 }
 
 export default CafeWallDemo;
-
