+++ app/experiments/hermann-grid/HermannGrid.tsx
