+++ app/routes.ts
@@ -1,11 +1,11 @@
 import { type RouteConfig, index, route } from "@react-router/dev/routes";
 
 export default [
-  index("routes/home.tsx"),
-  route("optical-illusion", "routes/optical-illusion.tsx"),
+  index("routes/home/home.tsx"),
+  route("optical-illusion", "routes/optical-illusion/optical-illusion.tsx"),
   route(
     "motion-induced-blindness",
-    "routes/motion-induced-blindness.tsx"
+    "routes/motion-induced-blindness/motion-induced-blindness.tsx"
   ),
-  route("hermann-grid", "routes/hermann-grid.tsx"),
+  route("hermann-grid", "routes/hermann-grid/hermann-grid.tsx"),
 ] satisfies RouteConfig;
