+++ app/routes/motion-induced-blindness/motion-induced-blindness.test.tsx
@@ -14,4 +14,3 @@ describe('Motion-Induced Blindness route', () => {
     ).toBeInTheDocument();
   });
 });
-
